+++ src/trx/game/lara/breath.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <trx/game/items.h>
+
+void Lara_Breath_Control(const ITEM *lara_item);