+++ docs/CHANGELOG.md
@@ -105,6 +105,7 @@
 - added sprite-based shadows
 - added footprints
 - added surface-based step sounds
+- added cold breath effects
 - added monochrome inventory backgrounds
 - added high-resolutions 16:9 and 4:3 loading screens
 - added high-resolutions 16:9 and 4:3 title and game end screens